Fix "DELETE /api/boards/[id]" api

+++ pages/api/boards/[id].ts
@@ -97,6 +97,62 @@ const handler = async (req: NextApiRequest, res: NextApiResponse<Content>) => {
         id: string;
       };
 
+      // Get lists that board has.
+      const { data: lists, error: listsError } = await supabase
+        .from("lists")
+        .select("*")
+        .eq("board_id", id)
+        .order("index");
+
+      if (listsError) {
+        throw listsError;
+      }
+
+      for (const list of lists) {
+        // Get cards that list has.
+        const { data: cards, error: cardsError } = await supabase
+          .from("cards")
+          .select("*")
+          .eq("list_id", list.id)
+          .order("index");
+
+        if (cardsError) {
+          throw cardsError;
+        }
+
+        // Delete checks that card has.
+        for (const card of cards) {
+          const { error } = await supabase
+            .from("checks")
+            .delete()
+            .eq("card_id", card.id);
+
+          if (error) {
+            throw error;
+          }
+        }
+
+        // Delete cards that list has.
+        const { error: deletedCardsError } = await supabase
+          .from("cards")
+          .delete()
+          .eq("list_id", list.id);
+
+        if (deletedCardsError) {
+          throw deletedCardsError;
+        }
+      }
+
+      // Delete lists that board has.
+      const { error: deletedListsError } = await supabase
+        .from("lists")
+        .delete()
+        .eq("board_id", id);
+
+      if (deletedListsError) {
+        throw deletedListsError;
+      }
+
       const { data: boards, error: boardsError } = await supabase
         .from("boards")
         .delete()
